requirement.txt
numpy
pandas
openpyxl
pycryptodome
requests
pymysql
beautifulsoup4
lxml
Flask-SQLAlchemy
flask-caching
PyMySQL
cryptography
WTForms
authlib
wtforms
flask-httpauth
openpyxl
flask-Cors